+++ meteor/server/api/snapshot.ts
@@ -583,8 +583,6 @@ export function restoreFromRundownPlaylistSnapshot(
 	const playlistId = (snapshot.playlist._id = getRandomId())
 	snapshot.playlist.restoredFromSnapshotId = snapshot.playlistId
 	delete snapshot.playlist.activationId
-	snapshot.playlist.currentPartInstanceId = null
-	snapshot.playlist.nextPartInstanceId = null
 
 	snapshot.rundowns.forEach((rd) => {
 		if (!rd.orphaned) {
@@ -679,6 +677,22 @@ export function restoreFromRundownPlaylistSnapshot(
 		}
 	})
 
+	if (snapshot.playlist.currentPartInstanceId) {
+		snapshot.playlist.currentPartInstanceId =
+			partInstanceIdMap[unprotectString(snapshot.playlist.currentPartInstanceId)] ||
+			snapshot.playlist.currentPartInstanceId
+	}
+	if (snapshot.playlist.nextPartInstanceId) {
+		snapshot.playlist.nextPartInstanceId =
+			partInstanceIdMap[unprotectString(snapshot.playlist.nextPartInstanceId)] ||
+			snapshot.playlist.nextPartInstanceId
+	}
+	if (snapshot.playlist.previousPartInstanceId) {
+		snapshot.playlist.previousPartInstanceId =
+			partInstanceIdMap[unprotectString(snapshot.playlist.previousPartInstanceId)] ||
+			snapshot.playlist.previousPartInstanceId
+	}
+
 	const rundownIds = snapshot.rundowns.map((r) => r._id)
 
 	// Apply the updates of any properties to any document
@@ -688,8 +702,8 @@ export function restoreFromRundownPlaylistSnapshot(
 			rundownId?: RundownId
 			partId?: PartId
 			segmentId?: SegmentId
-			part?: T
-			piece?: T
+			part?: unknown
+			piece?: unknown
 		}
 	>(objs: undefined | T[], updateId: boolean): T[] {
 		const updateIds = (obj: T) => {
@@ -709,10 +723,10 @@ export function restoreFromRundownPlaylistSnapshot(
 			}
 
 			if (obj.part) {
-				updateIds(obj.part)
+				updateIds(obj.part as any)
 			}
 			if (obj.piece) {
-				updateIds(obj.piece)
+				updateIds(obj.piece as any)
 			}
 
 			return obj
@@ -737,9 +751,9 @@ export function restoreFromRundownPlaylistSnapshot(
 	)
 	saveIntoDb(Segments, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.segments, false))
 	saveIntoDb(Parts, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.parts, false))
-	saveIntoDb(PartInstances, { rundownId: { $in: rundownIds } }, snapshot.partInstances)
+	saveIntoDb(PartInstances, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.partInstances, false))
 	saveIntoDb(Pieces, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.pieces, false))
-	saveIntoDb(PieceInstances, { rundownId: { $in: rundownIds } }, snapshot.pieceInstances)
+	saveIntoDb(PieceInstances, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.pieceInstances, false))
 	saveIntoDb(AdLibPieces, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.adLibPieces, true))
 	saveIntoDb(AdLibActions, { rundownId: { $in: rundownIds } }, updateItemIds(snapshot.adLibActions, true))
 	saveIntoDb(